fix: address PR #39 review feedback (round 2)
- docker-compose.prod.yml: pin default image tag to 0.27.0 instead of latest
- service.rs: add 30s timeout to command_output (prevents journalctl stall)
- service.rs: reject hostname+service combo in incident (journal has no host filter)
- cli.rs: surface dropped_lines in human-readable service-log output
- main.rs: add incident to HTTP-flags error command list
- mcp/tools.rs: document exclude_facility, received_from, received_to in syslog search help
- doctor.rs: move inline tests to sidecar doctor_tests.rs per repo layout rule
- CLAUDE.md: update version 0.25.3→0.27.0, add doctor.rs and checkpoint.rs, expand CLI table
