+++ src/cli.rs
@@ -2897,6 +2897,9 @@ fn print_service_logs_response(report: &ServiceLogsResponse, json: bool) -> Resu
         let message = entry.message.as_deref().unwrap_or("");
         println!("{timestamp} {ident}: {message}");
     }
+    if report.dropped_lines > 0 {
+        eprintln!("warning: {} malformed journal line(s) dropped", report.dropped_lines);
+    }
     Ok(())
 }
